return error when tenant with articles is removed

+++ src/SWP/Bundle/CoreBundle/Controller/TenantController.php
@@ -94,9 +94,12 @@ class TenantController extends FOSRestController
         $repository = $this->getTenantRepository();
         $tenant = $this->findOr404($code);
 
-//        $articlesRepository = $this->get('swp.repository.article');
-//        $existingArticles = $articlesRepository->findAll();
-        // todo check for articles
+        $articlesRepository = $this->get('swp.repository.article');
+        $existingArticles = $articlesRepository->findAll();
+
+        if (0 !== \count($existingArticles)) {
+            throw new ConflictHttpException('This tenant have articles attached to it.');
+        }
 
         $repository->remove($tenant);
